Move PgContextPlugin into @dataplan/pg
